rename to live_camera_config
