+++ pyxy3d/gui/live_camera_config/camera_summary_widget.py
